+++ repka/api.py
@@ -3,8 +3,3 @@ from repka.repositories.aiopg_ import AiopgRepository as BaseRepository
 from repka.repositories.aiopg_ import AiopgRepository
 from repka.repositories.fake import FakeRepo
 from repka.repositories.base import IdModel
-
-try:
-    from repka.repositories.databases_ import DatabasesRepository
-except ImportError:
-    pass