+++ repka/utils.py
@@ -1,6 +1,6 @@
 import json
 from contextlib import asynccontextmanager
-from typing import Sequence, Dict, Set, Union
+from typing import Sequence, Dict, Set, Union, TypeVar, Tuple, AsyncIterator, Iterator
 
 from aiopg.sa import SAConnection, create_engine
 from pydantic import BaseModel
@@ -43,3 +43,21 @@ async def create_async_db_connection(db_url: str) -> SAConnection:
 
 def is_field_equal_to_default(entity: BaseModel, field_name: str) -> bool:
     return getattr(entity, field_name) == entity.__fields__[field_name].default
+
+
+T = TypeVar("T")
+F = TypeVar("F")
+
+
+async def mixed_zip(first: Iterator[T], second: AsyncIterator[F]) -> AsyncIterator[Tuple[T, F]]:
+    """
+    A zip-like generator in which the first arg is a synchronous iterator, while the second one
+    is an async iterator.
+    """
+    first_iter = iter(first)
+    async for s in second:
+        try:
+            f = next(first_iter)
+        except StopIteration:
+            break
+        yield f, s